+++ app/src/main/java/com/example/eddy/basetrackerpsyegb/DB/RunMetrics.kt
@@ -9,7 +9,7 @@ import java.util.*
 data class RunMetrics(@PrimaryKey(autoGenerate = true) var id : Int,
                       @ColumnInfo(name = "startTime") var startTime : Long,
                       @ColumnInfo(name = "endTime") var endTime : Long,
-                      @ColumnInfo(name = "totalDistance") var totalDistance :Long){
+                      @ColumnInfo(name = "totalDistance") var totalDistance :Float){
     companion object{
         const val ID ="id"
         const val START_TIME = "startTime"
@@ -19,7 +19,7 @@ data class RunMetrics(@PrimaryKey(autoGenerate = true) var id : Int,
 
 
 
-    constructor():this(id = 0 , startTime = 0L, endTime = 0L, totalDistance = 0L)
+    constructor():this(id = 0 , startTime = 0L, endTime = 0L, totalDistance = 0F)
 
     override fun toString(): String {
         return "RunMetrics(id=$id, startTime=$startTime, endTime=$endTime, totalDistance=$totalDistance)"
